server/requirements.txt
beautifulsoup4==4.12.3
fake_useragent==1.5.1
Flask==3.0.3
groq==0.9.0
pinecone==5.0.0
pymongo==4.8.0
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.23.1
sentence_transformers==3.0.0
torch==2.3.1
webdriver_manager==4.0.2
